requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.1
Flask-Cors==2.1.2
progressbar==2.3
Flask-Bcrypt==0.7.1
PyJWT==1.5.3